pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "pillow>=11.1.0",
    "requests>=2.32.3",
    "selenium>=4.29.0",
    "trafilatura>=2.0.0",
]
